// Client/types/evaluation.ts
export interface Evaluation {
    id: number;
    color_size_recognition: number;
    belongings_memory: number;
    task_completion: number;
    application_id: number;
    total_mark: number;
    letter_number_distinction: number;
    stimuli_discrimination: number;
    auditory_memory: number;
    quick_responses: number;
    sustained_attention: number;
    environmental_perception: number;
    quick_comprehension: number;
    math_problem_solving: number,
    quranic_verses_recall: number
    first_time_attention: number;
    focus_on_significant_stimuli: number;
    createdAt: string;
    updatedAt: string;
  }